fix: use shared function
