+++ packages/@lwc/rollup-plugin/src/__tests__/fixtures/expected_compat_config_simple_app.js
@@ -85,9 +85,12 @@
 
   var __concat = Proxy.concat;
 
-  var cachedStylesheet;
-
-  function generateCss(hostSelector, shadowSelector, nativeShadow) {
+  function generateCss(
+    hostSelector,
+    shadowSelector,
+    nativeShadow,
+    hasAdoptedStyleSheets
+  ) {
     return nativeShadow
       ? ":host {color: var(--lwc-my-color);}"
       : __callKey1(
@@ -97,29 +100,7 @@
         );
   }
 
-  function stylesheet(
-    hostSelector,
-    shadowSelector,
-    nativeShadow,
-    hasAdoptedStyleSheets
-  ) {
-    if (nativeShadow && hasAdoptedStyleSheets) {
-      if (!cachedStylesheet) {
-        cachedStylesheet = new CSSStyleSheet();
-
-        __callKey1(
-          cachedStylesheet,
-          "replaceSync",
-          generateCss(hostSelector, shadowSelector, nativeShadow)
-        );
-      }
-
-      return cachedStylesheet; // fast path
-    }
-
-    return generateCss(hostSelector, shadowSelector, nativeShadow);
-  }
-
+  var stylesheet = lwc.createCachingCssGenerator(generateCss);
   var _implicitStylesheets = [stylesheet];
 
   function tmpl$1($api, $cmp, $slotset, $ctx) {
