+++ packages/@lwc/style-compiler/src/__tests__/fixtures/media-query-with-host/expected.js
@@ -1,19 +1,11 @@
 
 
-var cachedStylesheet
+import { createCachingCssGenerator } from 'lwc';
 
-function generateCss(hostSelector, shadowSelector, nativeShadow) {
+function generateCss(hostSelector, shadowSelector, nativeShadow, hasAdoptedStyleSheets) {
   return ["@media screen and (max-width: 768px) {", (nativeShadow ? ":host {width: calc(50% - 1rem);}" : [hostSelector, " {width: calc(50% - 1rem);}"].join('')), "}"].join('');
 }
 
-function stylesheet(hostSelector, shadowSelector, nativeShadow, hasAdoptedStyleSheets) {
-  if (nativeShadow && hasAdoptedStyleSheets) {
-    if (!cachedStylesheet) {
-      cachedStylesheet = new CSSStyleSheet();
-      cachedStylesheet.replaceSync(generateCss(hostSelector, shadowSelector, nativeShadow));
-    }
-    return cachedStylesheet; // fast path
-  }
-  return generateCss(hostSelector, shadowSelector, nativeShadow);
-}
+var stylesheet = createCachingCssGenerator(generateCss);
+
 export default [stylesheet];
