+++ packages/@lwc/style-compiler/src/__tests__/fixtures/multi-line/expected.js
@@ -1,19 +1,11 @@
 
 
-var cachedStylesheet
+import { createCachingCssGenerator } from 'lwc';
 
-function generateCss(hostSelector, shadowSelector, nativeShadow) {
+function generateCss(hostSelector, shadowSelector, nativeShadow, hasAdoptedStyleSheets) {
   return [(nativeShadow ? [":host(.selected) button", shadowSelector, " {color: #0070d2;}"].join('') : [hostSelector, ".selected button", shadowSelector, " {color: #0070d2;}"].join('')), (nativeShadow ? [":host:not(.selected) button", shadowSelector, " {color: #6d6d70;transition: all .2s ease-in-out;}"].join('') : [hostSelector, ":not(.selected) button", shadowSelector, " {color: #6d6d70;transition: all .2s ease-in-out;}"].join('')), (nativeShadow ? [":host:not(.selected):hover button", shadowSelector, ",:host button:focus", shadowSelector, " {color: #005fb2;transform: scale(1.2);}"].join('') : [hostSelector, ":not(.selected):hover button", shadowSelector, ",", hostSelector, " button:focus", shadowSelector, " {color: #005fb2;transform: scale(1.2);}"].join('')), "button:focus", shadowSelector, " {outline: none;box-shadow: none;background-color: #f3f2f2;}"].join('');
 }
 
-function stylesheet(hostSelector, shadowSelector, nativeShadow, hasAdoptedStyleSheets) {
-  if (nativeShadow && hasAdoptedStyleSheets) {
-    if (!cachedStylesheet) {
-      cachedStylesheet = new CSSStyleSheet();
-      cachedStylesheet.replaceSync(generateCss(hostSelector, shadowSelector, nativeShadow));
-    }
-    return cachedStylesheet; // fast path
-  }
-  return generateCss(hostSelector, shadowSelector, nativeShadow);
-}
+var stylesheet = createCachingCssGenerator(generateCss);
+
 export default [stylesheet];
